modified little code
