+++ src/components/History.tsx
@@ -14,8 +14,7 @@ import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
 import Paper from '@mui/material/Paper';
 
-import {ShopContext} from '../store';
-import {Order} from '../data';
+import {ShopContext,Order} from '../store';
 
 const StyledTableCell = styled(TableCell)(({theme}) => ({
   [`&.${tableCellClasses.head}`]: {
